tests/wpt/tests/navigation-api/navigate-event/navigate-window-open-self.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  navigation.onnavigate = t.step_func_done(e => {
    assert_equals(e.navigationType, "push");
    assert_true(e.cancelable);
    assert_true(e.canIntercept);
    assert_false(e.userInitiated);
    assert_true(e.hashChange);
    assert_equals(e.downloadRequest, null);
    assert_equals(new URL(e.destination.url).hash, "#1");
    assert_true(e.destination.sameDocument);
    assert_equals(e.destination.key, "");
    assert_equals(e.destination.id, "");
    assert_equals(e.destination.index, -1);
    assert_equals(e.formData, null);
    assert_equals(e.sourceElement, null);
    e.preventDefault();
  });
  window.onload = t.step_func(() => window.open("#1", "_self"));
}, "window.open() fires navigate event when targeting self");
</script>
